requirements.txt
# Django Core
Django==5.1.4
djangorestframework==3.15.2

# Database (PostgreSQL)
psycopg2-binary==2.9.10

# AI & Image Processing
# Using the updated 2025 Google Gen AI SDK
google-genai==1.0.0
Pillow==11.0.0

# Environment Variables
python-dotenv==1.0.1